requirements.txt
Django
djangorestframework
imutils
scipy
dlib
opencv-contrib-python